requirements.txt
eventlet!=0.18.3,!=0.20.1,>=0.18.2 # MIT
netaddr>=0.7.18 # BSD
SQLAlchemy>=1.4.23 # MIT
alembic>=1.6.5 # MIT
neutron-lib>=3.6.1 # Apache-2.0
os-ken >= 0.3.0 # Apache-2.0
oslo.config>=5.2.0 # Apache-2.0
oslo.db>=4.37.0 # Apache-2.0
oslo.log>=3.36.0 # Apache-2.0
oslo.messaging>=5.29.0 # Apache-2.0
oslo.service!=1.28.1,>=1.24.0 # Apache-2.0
oslo.utils>=3.33.0 # Apache-2.0
oslo.privsep>=1.32.0 # Apache-2.0
pyroute2>=0.7.2;sys_platform!='win32' # Apache-2.0 (+ dual licensed GPL2)
neutron>=23.0.0.0b2 # Apache-2.0
pyzmq>=14.3.1 # LGPL+BSD

ruamel.yaml
pydantic

# The comment below indicates this project repo is current with neutron-lib
# and should receive neutron-lib consumption patches as they are released
# in neutron-lib. It also implies the project will stay current with TC
# and infra initiatives ensuring consumption patches can land.
# neutron-lib-current
